Prevent client events from changing due to reuse of the same object in Keycloak
Keycloak uses an event builder to construct their events.
In some rare cases it can happen that the same builder gets reused inside the same transaction.
Due to the design of the builder, it will perform changes on the same object regardless if the event already got build (emitted) or not.

This behaviour is notable when signing in with an external identity provider, instead of pushing a REGISTER and LOGIN event it will publish the LOGIN event twice.